+++ src/N98/Magento/Command/Eav/Attribute/AbstractAttributeCommand.php
@@ -4,15 +4,16 @@ namespace N98\Magento\Command\Eav\Attribute;
 
 use N98\Magento\Command\AbstractMagentoCommand;
 use Magento\Eav\Model\Config;
+use Magento\Eav\Model\Entity\Attribute\AbstractAttribute;
 
 abstract class AbstractAttributeCommand extends AbstractMagentoCommand
 {
     /**
      * Gets an attribute model
-     * 
-     * @param  string $entityType
-     * @param  string $attributeCode
-     * @return Magento\Eav\Model\Entity\Attribute\AbstractAttribute
+     *
+     * @param string $entityType
+     * @param string $attributeCode
+     * @return AbstractAttribute
      */
     public function getAttribute($entityType, $attributeCode)
     {
@@ -20,5 +21,5 @@ abstract class AbstractAttributeCommand extends AbstractMagentoCommand
             ->getObjectManager()
             ->get(Config::class)
             ->getAttribute($entityType, $attributeCode);
-    }    
+    }
 }